♻️ Refactor: Composable 구조 분리
refactor: Composable 구조 분리 및 MainActivity 개선
